moved checkbox review page component to shared dir
