+++ src/applications/ivc-champva/10-10D/tests/components/CustomCheckboxRadioReviewPage.unit.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { CustomCheckboxRadioReviewPage } from '../../components/CustomCheckboxRadioReviewPage';
+import { CustomCheckboxRadioReviewPage } from '../../../shared/components/CustomCheckboxRadioReviewPage';
 import { testComponentRender } from '../../../shared/tests/pages/pageTests.spec';
 import mockData from '../fixtures/data/test-data.json';
 import { generateOptions } from '../../pages/ApplicantMedicareStatusPage';
